Fixed error checking for test file downloading

+++ test/end_to_end_test.py
@@ -73,7 +73,8 @@ print 'creating dataruns...'
 datarun_ids = []
 for ds in DATASETS:
     # remove existing and redownload data to ensure clasifiers are generated
-    os.remove(DATA_DIR+ds)
+    if ds in os.listdir(DATA_DIR):
+        os.remove(DATA_DIR+ds)
     run_config.train_path = download_file_url(S3_BUCKET+ds, DATA_DIR)
     dataset = enter_dataset(db, run_config, aws_config)
     datarun_ids.append(enter_datarun(sql_config, run_config, aws_config))
